[UPDATE]: updated startServer to change the name and adding the files dynamically
